category issue and wishlist issue fixed

+++ app/Http/Controllers/Api/WishLishController.php
@@ -7,8 +7,10 @@ use App\Helpers\ResponseBuilder;
 use App\Traits\ResponceTrait;
 use App\Traits\RestControllerTrait;
 use App\WishLish;
+use App\VendorProduct;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Resources\WishListResource;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
@@ -40,35 +42,19 @@ class WishLishController extends Controller
     }
     public function index(Request $request)
     {
-        $user = Auth::guard('api')->user();
-        $wishlists = $user->wishlist()->with('vendorProduct','vendorProduct.Product.image')->get();
-        $data= $this->wishLish->where(['user_id'=>Auth::guard('api')->user()->id])->where(['zone_id'=>Auth::guard('api')->user()->zone_id])->has('vendorProduct')->has('vendorProduct.Product')->with(['vendorProduct.Product.image']);
-       
-        $result= [];
-        foreach ($data as $rec){
-            $rec['vendor_product']['product']['image']=$rec['vendor_product']['product']['image']['name'];
-            $result[]= $rec;
+        try{
+            $user = Auth::guard('api')->user();
+            $wishlists = $user->wishlist()->whereHas('vendorProduct')->with(['vendorProduct' => function($q){
+                $q->whereNull('deleted_at');
+            }],'vendorProduct.Product.image')->where('zone_id', $user->zone_id)->paginate(20);
+            $this->response->wish_list = WishListResource::collection($wishlists);
+            $this->response->wish_list_count = $wishlists->count();
+            return ResponseBuilder::successWithPagination($wishlists, $this->response, $this->successStatus);
+        }catch (\Exception $e){
+            return ResponseBuilder::error($e->getMessage(), $this->errorStatus);
         }
-
-        $response = [
-            'error'=>false,
-            'code' => 0,
-            'wish_list' => $data,
-            'wish_list_count' => count($result),
-            'message'=>trans('site.success'),
-        ];
-
-        return response()->json($response, 200);
-
-
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(),$this->wishLish->rules($this->method),$this->wishLish->messages($this->method));
@@ -78,15 +64,20 @@ class WishLishController extends Controller
         }else{
             try {
                 $user = Auth::guard('api')->user();
+                //get vendor product
+                $vendor_product = VendorProduct::find($request->vendor_product_id);
+                if(!$vendor_product){
+                    return ResponseBuilder::error("Vendor product not found", $this->validationStatus);
+                }
                 //get user wishlist
-                $wishlist = $user->wishlist()->where(['vendor_product_id'=>$request->vendor_product_id])->first();
+                $wishlist = $user->wishlist()->where(['vendor_product_id'=>$request->vendor_product_id, 'zone_id' => $user->zone_id])->first();
                 if($wishlist){
                     //remove from wishlist
                     $wishlist->delete();
                     return ResponseBuilder::success(null, trans('site.wishlist_delete'), $this->successStatus);
                 }else{
                     //add to wishlist
-                    $user->wishlist()->create(['vendor_product_id'=>$request->vendor_product_id]);
+                    $user->wishlist()->create(['vendor_product_id'=>$request->vendor_product_id, 'zone_id'=>$user->zone_id]);
                     return ResponseBuilder::success(null,trans('site.wishlist_create'), $this->successStatus);
                 }
             } catch (\Exception $e) {
